Refactor: move stack-graph builder into a bindings module
