+++ crates/codegen/runtime/cargo/src/runtime/bindings/stack_graph/functions.rs
